requirements.txt
fastapi
uvicorn
requests
pydub
jinja2
python-multipart
openai-whisper